Added a "glob" property to the files config that represents the un-expanded globs for each file pattern
This is useful for the watch command, which needs to
continuously re-expand the paths. We can also use the
globs with any task that we know performs expansion
itself.

How many tasks perform expansion at task-time? If it's
a lot, we may be able to drastically reduce the number
of lineman configure task executions