when lineman is called and no command matches, it will try calling grunt with those commands as if it was given an arbitrary task name.
This wasn't added for funsies, because the 
watch task changed how it spawns grunt in such a
way that it will inadvertently spawn lineman on
a watch match. As a result it seemed easiest to just
quack like grunt.

HOWEVER this seems to be significantly slower from
the user's experience. which is totally not cool.