Removes testem from the user's package.json
This one's a little more fun than the grunt
and grunt-contrib commits. Not only does it pull
testem out of the user's package.json, but it
will allow users to install their own testem 
if lineman ever falls behind and the user's
testem runner will win if it exists. 

I imagine this will be very handy b/c of how
fast-moving testem is. @davemo should be pleased.