Added coverage for adding a CoffeeScript file
This is actually getting kinda nifty. Starts lineman
run in a thread, then adds a file and will hammer it
for up to 5 seconds until the CoffeeScript function
shows up in the source and works.

At this point need to start managing the threads
responsibly. The two tests will each open separate
ones.