No longer require lineman to be locally installed
If lineman is locally installed (and it will be as
soon as the user runs `npm install`), then that version
will always win, but mercifully lineman will now
function just as well by being completely global.

This brings with it a few changes:

* lineman new's --skip-install option is now replaced
by --install.
* lineman's cli will now own the responsibility of
figuring out from where the user should be requiring
lineman, and it's setting that location (either an
absolute path to global lineman or just `lineman`)
to an environment variable called LINEMAN_MAIN. 
**This will require a change to people's archetypes
to pick up, but won't be a regression since if they're
working today they've already got Lineman installed locally**