This solves apiProxy eating everything
But also means people will have to provide a prefix for api endpoints.
I'm torn on this one, I know we used to have a hardcoded prefix in here
but if people are creating a single page app and using HTML5 pushState
all the examples I've found indicate that a prefix is used.