fixes #46 (finally!)
Reproduced this with a super simple POST request,
was much easier to reproduce once we knew that
ordering was the issue.

``` ruby
require 'sinatra'

post '/hi' do
  puts params
end
```

Ran that server, connected the bodyParser at the
top of the express setup, then tried this from
console:

``` javascript
$.post('/hi', {wtf: true})
```

And sure enough, it hung forever.

Moved the bodyParser beneath the proxying setup
and it worked fine.

yay!