Adds new config property "server.relativeUrlRoot"
This property, when set in the config as, say,

```
server: {
  relativeUrlRoot: "/foo"
}
```

Will mount the entire dev server onto that relative path. You'll
see this output when you start the server:

```
Mounting application at path '/foo'.
```

Because this won't be a commonly used feature, you can expect just as 
many quirks as you do with the Ruby on Rails' equivalent. :-)

YMMV, Fixes #188, YOLO