Add customUrlPath config param for server task
We would like to be able to emulate our production paths as closely as
possible in development. What we want is to mount the app at something
like "/development/engagement", but still hit the back-end API at
"/development/api/..."

Using the 'relativeUrlRoot', I have not been able to figure out a way to
do this. It appears that the relativeUrlRoot option is used both by the
api proxy *and* by the express server, and therefore it will only work if
we want the same relative path for both servers.

I am aware that with my complete lack of "node/express"-fu, this is quite
possibly the WRONG way to do this, so I'm asking for input. Is this
reasonable? Or am I completely missing something obvious in the
application config options?