Allow users to specify their watch module
Lineman defaults don't change, but users can define:

```
watch: {
  options: {
    module: "grunt-contrib-watch"
  }
}
```

And install `npm i --save-dev grunt-contrib-watch` locally, and that 
one will be used for the watch task.

The goal is to give a path for people who outgrow our 
grunt-watch-nospawn for feature's or project size, and are willing to 
sacrifice its modest speed advantage without creating a lot of friction.